Adding options for non-interactive stream selection.
New option --auto selects the 'best' stream available without
user intervention. Options --max-quality and ---subs can be used
to constrain the selected value to a max quality and/or a given
kind of subs.

The max-quality is a hack, it requires the argument to be
formatted the same way as the given source e. g. '850 kbps' or
'768x512'. Proper solution would be to add something to services
which decodes generic values like 'low', 'medium', 'high' and
'max' for the given service.